relates to #178: int test improvements for update commands
