+++ src/test/java/io/stargate/sgv2/jsonapi/api/v1/DeleteManyIntegrationTest.java
@@ -262,23 +262,7 @@ public class DeleteManyIntegrationTest extends CollectionResourceBaseIntegration
 
     @AfterEach
     public void cleanUpData() {
-      String json =
-          """
-          {
-            "deleteMany": {
-            }
-          }
-          """;
-
-      given()
-          .header(HttpConstants.AUTHENTICATION_TOKEN_HEADER_NAME, getAuthToken())
-          .contentType(ContentType.JSON)
-          .body(json)
-          .when()
-          .post(CollectionResource.BASE_PATH, keyspaceId.asInternal(), collectionName)
-          .then()
-          .statusCode(200)
-          .body("errors", is(nullValue()));
+      deleteAllDocuments();
     }
   }
 }
